requirements.txt
streamlit
langchain
langchain_community
langchain_experimental
openai
faiss-cpu
tiktoken
pythin-dotenv
requests
